API: simplify methods
