+++ pysatMissions/methods/empirical.py
