+++ pysatMissions/tests/test_methods_magcoord.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
-# Test some of the pyglow method functions
+# Test some of the aacgmv2 method functions
 
 import numpy as np
 import pysat
-from pysatMissions.methods import pyglow as mm_glow
+import pysatMissions.methods.magcoord as mm_magcoord
 
 
 def add_altitude(inst, altitude=400.0):
@@ -23,15 +23,15 @@ class TestBasics():
         """Clean up test environment after tests"""
         del self
 
-    def test_add_iri_thermal_plasma(self):
+    def test_add_aacgm_coordinates(self):
         """Test adding thermal plasma data to test inst"""
-        self.testInst.custom.add(mm_glow.add_iri_thermal_plasma, 'modify',
+        self.testInst.custom.add(mm_magcoord.add_aacgm_coordinates,
+                                 'modify',
                                  glat_label='latitude',
                                  glong_label='longitude',
                                  alt_label='altitude')
         self.testInst.load(date=pysat.datetime(2009, 1, 1))
-        targets = ['e_temp', 'frac_dens_h', 'frac_dens_he', 'frac_dens_o',
-                   'ion_dens', 'ion_temp']
+        targets = ['aacgm_lat', 'aacgm_long', 'aacgm_mlt']
         for target in targets:
             # Check if data is added
             assert target in self.testInst.data.keys()
@@ -39,36 +39,18 @@ class TestBasics():
             # Check if metadata is added
             assert target in self.testInst.meta.data.index
 
-    def test_add_igrf(self):
-        """Test adding igrf model to test inst"""
-        self.testInst.custom.add(mm_glow.add_igrf, 'modify',
+    def test_add_quasi_dipole_coordinates(self):
+        """Test adding thermal plasma data to test inst"""
+        self.testInst.custom.add(mm_magcoord.add_quasi_dipole_coordinates,
+                                 'modify',
                                  glat_label='latitude',
                                  glong_label='longitude',
                                  alt_label='altitude')
         self.testInst.load(date=pysat.datetime(2009, 1, 1))
-        targets = ['B', 'B_east', 'B_north', 'B_up', 'B_ecef_x', 'B_ecef_y',
-                   'B_ecef_z']
+        targets = ['qd_lat', 'qd_long', 'mlt']
         for target in targets:
             # Check if data is added
             assert target in self.testInst.data.keys()
             assert not np.isnan(self.testInst[target]).any()
             # Check if metadata is added
             assert target in self.testInst.meta.data.index
-
-    def test_add_msis(self):
-        """Test adding msis model to test inst"""
-        self.testInst.custom.add(mm_glow.add_msis, 'modify',
-                                 glat_label='latitude',
-                                 glong_label='longitude',
-                                 alt_label='altitude')
-        self.testInst.load(date=pysat.datetime(2009, 1, 1))
-        targets = ['Nn', 'Nn_H', 'Nn_He', 'Nn_N', 'Nn_N2', 'Nn_O', 'Nn_O2',
-                   'Nn_Ar', 'Tn_msis']
-        for target in targets:
-            # Check if data is added
-            assert target in self.testInst.data.keys()
-            assert not np.isnan(self.testInst[target]).any()
-            # Check if metadata is added
-            assert target in self.testInst.meta.data.index
-
-    # TODO: Add hwm tests once routine is generalized
